requirements.txt
aenum==3.0.0
aniso8601==9.0.1
attrs==20.3.0
blis==0.7.4
boto3==1.17.45
botocore==1.20.45
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.0.0/es_core_news_md-3.0.0-py3-none-any.whl
es-dep-news-trf @ https://github.com/explosion/spacy-models/releases/download/es_dep_news_trf-3.0.0/es_dep_news_trf-3.0.0-py3-none-any.whl
filelock==3.0.12
Flask==1.1.2
flask-restplus==0.13.0
ftfy==5.9
idna==2.10
importlib-metadata==3.10.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
MarkupSafe==1.1.1
murmurhash==1.0.5
numpy==1.20.2
packaging==20.9
pandas==1.2.3
pathy==0.4.0
pluggy==0.13.1
preshed==3.0.5
py==1.10.0
pydantic==1.7.3
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
s3transfer==0.3.6
sacremoses==0.0.44
six==1.15.0
smart-open==3.0.0
spacy==3.0.5
spacy-alignments==0.7.2
spacy-legacy==3.0.2
spacy-transformers==1.0.1
srsly==2.4.0
thinc==8.0.2
tokenizers==0.9.4
toml==0.10.2
torch==1.8.1
torchcontrib==0.0.2
tqdm==4.60.0
transformers==4.2.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==0.16.1
zipp==3.4.1